+++ packages/docs/src/pages/TestPage.tsx
@@ -2,7 +2,7 @@ import * as yup from 'yup'
 
 import { Button } from '@siakit/button'
 import { Footer } from '@siakit/footer'
-import { Form, RichTextInput, useForm } from '@siakit/form-unform'
+import { Form, SelectMulti, useForm } from '@siakit/form-unform'
 import { Flex } from '@siakit/layout'
 
 const testFormSchema = yup.object({
@@ -25,10 +25,23 @@ export function TestPage() {
     <Flex direction="column" gap padding>
       <Form ref={formRef} onSubmit={() => handleSubmit(submit)}>
         <Flex height={400}>
-          <RichTextInput
+          <SelectMulti
             name="description"
             label="Description"
             placeholder="Description"
+            options={[
+              {
+                value: 1,
+                label:
+                  'Lorem ipsum dolor sit amet consectetur, adipisicing elit. Blanditiis nesciunt, Lorem ipsum dolor sit amet consectetur, adipisicing elit. Blanditiis nesciunt',
+              },
+              {
+                value: 2,
+                label:
+                  'Lorem ipsum dolor sit amet consectetur, adipisicing elit. Blanditiis nesciunt, Lorem ipsum dolor sit amet consectetur, adipisicing elit. Blanditiis nesciunt',
+              },
+            ]}
+            closeMenuOnSelect={false}
           />
         </Flex>