+++ packages/form-unform/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @siakit/form-unform
 
+## 0.0.42
+
+### Patch Changes
+
+- fix select multi wrap text
+- Updated dependencies
+  - @siakit/form-components@0.0.22
+
 ## 0.0.41
 
 ### Patch Changes